Reorganize directory service for testing
