+++ services/directory-service/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod data;
+pub mod database;
+pub mod manager;
+pub mod rpc;
+pub mod startup;
