+++ services/directory-service/src/main.rs
@@ -2,13 +2,11 @@ mod data;
 mod database;
 mod manager;
 mod rpc;
+mod startup;
 
-use std::{path::PathBuf, sync::Arc};
 use structopt::StructOpt;
 
-use data::{
-    config::DirectoryConfig, context::DirectoryContext, error::DirectoryError, opt::DirectoryOpt,
-};
+use data::opt::DirectoryOpt;
 
 #[tokio::main]
 async fn main() {
@@ -16,36 +14,12 @@ async fn main() {
         .with_max_level(tracing::Level::INFO)
         .init();
 
-    let context = build_context(DirectoryOpt::from_args())
+    let config = startup::build_config(DirectoryOpt::from_args().config_path)
+        .expect("Failed to parse config");
+
+    let context = startup::build_context(config)
         .await
         .expect("Failed to generate context");
 
     rpc::server::start_server(context).await;
-}
-
-async fn build_context(
-    directory_opt: DirectoryOpt,
-) -> Result<Arc<DirectoryContext>, DirectoryError> {
-    let config = build_config(directory_opt.config_path)?;
-
-    Ok(Arc::new(data::context::DirectoryContext {
-        host_address: config.host_address,
-        database_pool: database::general::database_init(&config.database_path).await?,
-    }))
-}
-
-fn build_config(yaml_path: PathBuf) -> Result<DirectoryConfig, DirectoryError> {
-    match std::fs::read_to_string(yaml_path) {
-        Ok(yaml_string) => match serde_yaml::from_str(&yaml_string) {
-            Ok(config) => Ok(config),
-            Err(e) => Err(DirectoryError::ConfigError(format!(
-                "Config contents invalid: {}",
-                e
-            ))),
-        },
-        Err(e) => Err(DirectoryError::ConfigError(format!(
-            "Failed to open config file: {}",
-            e
-        ))),
-    }
 }
